add an example for other file types
Unfortunately, the very simple `write_cpio` helper function will only
write out an archive of files. Anything more complicated and you wind up
having to manually build out the record.

The "other-file-types" example does that to add an empty directory to
the archive as well as a symlink. For the empty directory, you just have
to set the mode file type to directory and then set the length to 0. For
a symlink, you set the mode file type to symlink, but then the
_contents_ of that file needs to be where the symlink will point.

Fixes #18